+++ app/code/community/Nosto/Tagging/Helper/Account.php
@@ -311,4 +311,23 @@ class Nosto_Tagging_Helper_Account extends Mage_Core_Helper_Abstract
 
         return true;
     }
+
+    /**
+     * Returns all store views that have Nosto intsalled
+     *
+     * @return Mage_Core_Model_Store[]
+     */
+    public function getAllStoreViewsWithNostoAccount()
+    {
+        $storesWithNosto = array();
+        /* @var Nosto_Tagging_Helper_Data $helper */
+        $helper = Mage::helper('nosto_tagging');
+        foreach ($helper->getAllStoreViews() as $store) {
+            if ($this->existsAndIsConnected($store)) {
+                $storesWithNosto[] = $store;
+            }
+        }
+
+        return $storesWithNosto;
+    }
 }