+++ app/code/community/Nosto/Tagging/Model/Resource/Index/Collection.php
@@ -1,9 +1,9 @@
 <?php
 /**
  * Magento
- *  
+ *
  * NOTICE OF LICENSE
- *  
+ *
  * This source file is subject to the Open Software License (OSL 3.0)
  * that is bundled with this package in the file LICENSE.txt.
  * It is also available through the world-wide-web at this URL:
@@ -11,13 +11,13 @@
  * If you did not receive a copy of the license and are unable to
  * obtain it through the world-wide-web, please send an email
  * to license@magentocommerce.com so we can send you a copy immediately.
- *  
+ *
  * DISCLAIMER
- *  
+ *
  * Do not edit or add to this file if you wish to upgrade Magento to newer
  * versions in the future. If you wish to customize Magento for your
  * needs please refer to http://www.magentocommerce.com for more information.
- *  
+ *
  * @category  Nosto
  * @package   Nosto_Tagging
  * @author    Nosto Solutions Ltd <magento@nosto.com>
@@ -26,30 +26,19 @@
  */
 
 /**
- * Meta data class which holds information about the current backend user.
+ * Index resource collection.
  *
  * @category Nosto
  * @package  Nosto_Tagging
  * @author   Nosto Solutions Ltd <magento@nosto.com>
  */
-class Nosto_Tagging_Model_Meta_User extends Nosto_Object_User
+class Nosto_Tagging_Model_Resource_Index_Collection extends Mage_Core_Model_Resource_Db_Collection_Abstract
 {
     /**
-     * Loads the user data from the active session.
-     *
-     * @return bool
+     * @inheritdoc
      */
-    public function loadData()
+    protected function _construct()
     {
-        /** @var Mage_Admin_Model_User $user */
-        /** @noinspection PhpUndefinedMethodInspection */
-        $user = Mage::getSingleton('admin/session')->getUser();
-        if ($user) {
-            $this->setFirstName($user->getFirstname());
-            $this->setLastName($user->getLastname());
-            $this->setEmail($user->getEmail());
-        }
-
-        return true;
+        $this->_init('nosto_tagging/index');
     }
 }